+++ tests/conftest.py
@@ -5,4 +5,9 @@ import pytest
 
 @pytest.fixture
 def example_path() -> Path:
-    return Path(__file__).parents[1] / "examples" / "un_afm.xyz"
+    return (
+        Path(__file__).parents[1]
+        / "examples"
+        / "afm_frame"
+        / "trajectory.xyz"
+    )